move spu smartmodule to controlplane only
